+++ snowddl/resolver/schema_read_role.py
@@ -4,7 +4,6 @@ from snowddl.blueprint import (
     Grant,
     RoleBlueprint,
     SchemaBlueprint,
-    SchemaObjectIdent,
     build_role_ident,
 )
 from snowddl.resolver.abc_role_resolver import AbstractRoleResolver, ObjectType
@@ -73,17 +72,3 @@ class SchemaReadRoleResolver(AbstractRoleResolver):
         )
 
         return bp
-
-    def grant_to_future_grant(self, grant: Grant):
-        if not grant.on.is_future_grant_supported:
-            return None
-
-        if isinstance(grant.name, SchemaObjectIdent):
-            return FutureGrant(
-                privilege=grant.privilege,
-                on_future=grant.on,
-                in_parent=ObjectType.SCHEMA,
-                name=grant.name.schema_full_name,
-            )
-
-        return None